requirements.txt
scikit-learn==0.22.1
pandas==0.24.0
numpy==1.17.4
tkcalendar==1.6.1
Pillow==7.0.0
xlwt==1.3.0
xlrd==1.2.0
